% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateExpectedEventTimes.R
\name{calculateExpectedEventTimes}
\alias{calculateExpectedEventTimes}
\title{Calculate Expected Event Time for Multiple Observations in Coxian Phase-Type Distribution}
\usage{
calculateExpectedEventTimes(
  model_object,
  new_observations,
  n_phases,
  upper_time = 10000,
  strata_by,
  file_path = NULL
)
}
\arguments{
\item{model_object}{An object containing the fitted model, including transition rates.}

\item{new_observations}{A dataframe containing the covariate values for new observations.}

\item{n_phases}{The total number of phases in the Coxian model.}

\item{upper_time}{Optional upper limit for numerical integration to approximate infinity. Defaults to 10000, which should be adjusted based on the distribution's characteristics.}

\item{strata_by}{A string specifying the column name in new_observations used for stratification, aligning with the model's structure.}

\item{file_path}{Optional file path to save each row's results to a CSV file. If not provided, results will not be saved to a file.}
}
\value{
The input dataframe with an additional `estimatedEventTime` column containing the expected event time for each observation.
}
\description{
Extends the calculateExpectedEventTime function to handle a dataframe of new observations,
adding an estimatedEventTime column to the dataframe with the expected event time for each observation.
If an error occurs during the calculation for any observation, the function sets the estimatedEventTime to "ERROR"
for that observation and continues processing the remaining observations.
}
\examples{
\dontrun{
  # Assume `model_object` is a pre-fitted model object
  # `new_observations` is a dataframe with multiple rows, each row having covariate values
  result_df <- calculateExpectedEventTimes(model_object, new_observations,
                    n_phases = 3, upper_time = 10000, strata_by = "phaseColumn", file_path = "results.csv")
  print(result_df)
}

}
