% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Cin.R
\name{compute_Cin}
\alias{compute_Cin}
\title{Compute the C_{i,n} term for the Coxian PDF}
\usage{
compute_Cin(lambda, mu, i, n)
}
\arguments{
\item{lambda}{A numeric vector of lambda values.}

\item{mu}{A numeric vector of mu values.}

\item{i}{An integer indicating the i-th term.}

\item{n}{An integer indicating the phase.}
}
\value{
A numeric value representing C_{i,n}.
}
\description{
This function computes the C_{i,n} term used in the Coxian phase-type distribution PDF.
}
\keyword{internal}
