% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pn.R
\name{compute_pn}
\alias{compute_pn}
\title{Compute the p_n term for the Coxian PDF}
\usage{
compute_pn(lambda, mu, n)
}
\arguments{
\item{lambda}{A numeric vector of lambda values.}

\item{mu}{A numeric vector of mu values.}

\item{n}{An integer indicating the phase for which p_n is to be computed.}
}
\value{
A numeric value representing p_n.
}
\description{
This function computes the p_n term used in the Coxian phase-type distribution PDF.
}
\keyword{internal}
