% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxianAic.R
\name{coxianAic}
\alias{coxianAic}
\title{Akaike Information Criterion for Coxian Distribution}
\usage{
coxianAic(data, lambda, mu)
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values for the Coxian distribution.}

\item{mu}{A numeric vector of mu values for the Coxian distribution.}
}
\value{
A numeric value representing the AIC.
}
\description{
Computes the Akaike Information Criterion (AIC) for a given set of data and Coxian distribution parameters.
}
\examples{
data_sample <- rexp(100, rate = 0.5)
lambda <- c(0.5, 0)
mu <- c(0.2, 0.4)
coxianAic(data_sample, lambda, mu)

}
