% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxianPdf.R
\name{coxianPdf}
\alias{coxianPdf}
\title{Coxian Phase-Type Distribution PDF}
\usage{
coxianPdf(t, lambda, mu)
}
\arguments{
\item{t}{A numeric value representing time.}

\item{lambda}{A numeric vector of lambda values.}

\item{mu}{A numeric vector of mu values.}
}
\value{
A numeric value representing the PDF value at time t.
}
\description{
Computes the probability density function (PDF) of a Coxian phase-type distribution for a given time value t.
}
\examples{
lambda <- c(0.5, 0.3, 0)
mu <- c(0.2, 0.4, 0.6)
t_value <- 1
pdf_value <- coxian_pdf(t_value, lambda, mu)
print(pdf_value)

}
