% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_transition_rates.R
\name{estimate_transition_rates}
\alias{estimate_transition_rates}
\title{Estimate Transition Rates}
\usage{
estimate_transition_rates(
  coxph_object,
  new_observation,
  n_phases = NULL,
  strata_by
)
}
\arguments{
\item{coxph_object}{A coxph object from the fitCoxPhModel function.}

\item{new_observation}{A data frame with the covariate values for the new observation.}

\item{n_phases}{The number of phases in the Coxian model.}
}
\value{
A list containing the estimated lambda and mu values for each phase.
}
\description{
Estimate Transition Rates
}
\examples{
\dontrun{
  # Assuming fit_result is a coxph object from the fitCoxPhModel function
  # and new_obs is a data frame with the covariate values for the new observation
  transition_rates <- estimate_transition_rates(fit_result, new_obs, n_phases = 3)
  print(transition_rates)
}

}
