% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCoxian.R
\name{fitCoxian}
\alias{fitCoxian}
\title{Fit Coxian Phase Type Distribution to Data}
\usage{
fitCoxian(data, n_phases)
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{n_phases}{An integer specifying the number of phases for the Coxian distribution.}
}
\value{
A list containing the optimized lambda and mu values, negative log-likelihood, sum of squared differences (SSD), Akaike Information Criterion (AIC), and Kolmogorov-Smirnov test results.
}
\description{
Fits a Coxian phase type distribution to the provided data based on the specified number of phases.
}
\examples{
data_sample <- rexp(100, rate = 0.5)
fit_results <- fitCoxianPH(data_sample, 2)

}
