% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_density.R
\name{interpolate_density}
\alias{interpolate_density}
\title{Interpolate Density from Kernel Density Estimation}
\usage{
interpolate_density(x, kde)
}
\arguments{
\item{x}{A numeric value or vector for which the density is to be interpolated.}

\item{kde}{A kernel density estimation object, typically obtained from the `density` function.}
}
\value{
A numeric value or vector representing the interpolated density at `x`.
}
\description{
This function interpolates the density value for a given point `x` based on
the provided kernel density estimation (KDE).
}
\examples{
data_sample <- rexp(100, rate = 0.5)
kde <- density(data_sample)
interpolate_density(0.5, kde)

}
