% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksTestCoxian.R
\name{ksTestCoxian}
\alias{ksTestCoxian}
\title{Kolmogorov-Smirnov Test for Coxian Distribution}
\usage{
ksTestCoxian(data, lambda, mu)
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values for the Coxian distribution.}

\item{mu}{A numeric vector of mu values for the Coxian distribution.}
}
\value{
A list with class "htest" representing the test results.
}
\description{
Performs the Kolmogorov-Smirnov test on data to compare it with a Coxian distribution.
}
\examples{
data_sample <- rexp(100, rate = 0.5)
lambda <- c(0.5, 0)
mu <- c(0.2, 0.4)
ksTestCoxian(data_sample, lambda, mu)

}
