% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadModelFromFile.R
\name{loadModelFromFile}
\alias{loadModelFromFile}
\title{Load Model from File}
\usage{
loadModelFromFile(filepath)
}
\arguments{
\item{filepath}{A string specifying the path to the file where the model object is saved. The file
should have been created using `saveRDS()`.}
}
\value{
The model object loaded from the specified file. If an error occurs during loading, `NULL`
is returned and an error message is printed to the console.
}
\description{
This function loads a model object from a specified file path. It is designed to work with models
that have been saved using `saveRDS()`. This is useful for scenarios where you need to persist
model objects on disk for later use, such as for prediction or further analysis.
}
\examples{
\dontrun{
  # Assuming you have previously saved a model to "model.rds"
  model_path <- "path/to/your_saved_model.rds"
  loaded_model <- loadModelFromFile(model_path)
  # Now `loaded_model` can be used for predictions or further analysis
}

}
