% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{optimize}
\alias{optimize}
\title{Optimize Parameters for Coxian Phase Type Distribution}
\usage{
optimize(data, init_lambda, init_mu, optim_method = "BFGS")
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{init_lambda}{A numeric vector of initial values for lambda.}

\item{init_mu}{A numeric vector of initial values for mu.}

\item{optim_method}{A character string specifying the optimization method to be used by `optim`.
Default is "BFGS".}
}
\value{
A list containing the optimized lambda and mu values.
}
\description{
This function optimizes the parameters lambda and mu of a Coxian phase type distribution
to fit the provided data. The optimization method can be controlled through the `optim_method` parameter.
}
\examples{
data_sample <- rexp(100, rate = 0.5)
init_lambda <- c(0.5, 0)
init_mu <- c(0.2, 0.4)
result <- optimize(data_sample, init_lambda, init_mu)

}
