% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogramPdf.R
\name{plotHistogramPdf}
\alias{plotHistogramPdf}
\title{Plot Histogram with Coxian PDF Overlay}
\usage{
plotHistogramPdf(
  data,
  lambda,
  mu,
  breaks = 30,
  main = "Fit of Coxian PH",
  xlab = "Value",
  ylab = "Density",
  curve_col = "blue",
  curve_lwd = 2
)
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values for the Coxian distribution.}

\item{mu}{A numeric vector of mu values for the Coxian distribution.}

\item{breaks}{The number of breaks for the histogram. Default is 30.}

\item{main}{The title of the plot. Default is "Fit of Coxian PH".}

\item{xlab}{The x-axis label. Default is "Value".}

\item{ylab}{The y-axis label. Default is "Density".}

\item{curve_col}{The color of the Coxian PDF curve. Default is "blue".}

\item{curve_lwd}{The line width of the Coxian PDF curve. Default is 2.}
}
\value{
A plot showing the histogram with the Coxian PDF overlay.
}
\description{
Plots a histogram of the observed data and overlays the Coxian phase type distribution's PDF.
}
\examples{
data_sample <- rexp(100, rate = 0.5)
lambda <- c(0.5, 0)
mu <- c(0.2, 0.4)
plotHistogramPdf(data_sample, lambda, mu, main = "Custom Title", curve_col = "red")

}
