% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Residuals Plot for Coxian Distribution}
\usage{
plotResiduals(data, lambda, mu)
}
\arguments{
\item{data}{A numeric vector of observed data points.}

\item{lambda}{A numeric vector of lambda values for the Coxian distribution.}

\item{mu}{A numeric vector of mu values for the Coxian distribution.}
}
\value{
A plot showing the residuals.
}
\description{
Plots the residuals of the observed data against the expected values from the Coxian distribution.
}
\examples{
data_sample <- rexp(100, rate = 0.5)
lambda <- c(0.5, 0)
mu <- c(0.2, 0.4)
plotResiduals(data_sample, lambda, mu)

}
