% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCoxianPdf.R
\name{predictCoxianPdf}
\alias{predictCoxianPdf}
\title{Predict Coxian PDF}
\usage{
predictCoxianPdf(
  model_object,
  new_observation,
  current_phase,
  current_time,
  strata_by
)
}
\arguments{
\item{model_object}{A list containing the trained model output.
The `transition_rates` must include lambda (transition rates between phases) and mu (absorption rates).}

\item{new_observation}{A dataframe containing the covariate values for the new observation.}

\item{current_phase}{An integer specifying the current phase of the process for the new observation.
It must be between 1 and the total number of phases in the model.}

\item{current_time}{A numeric value specifying the time at which to calculate the Coxian PDF.}

\item{strata_by}{A string specifying the column name in `new_observation` that represents the phase.}
}
\value{
A numeric value representing the calculated PDF of the Coxian phase-type model at the specified time.
}
\description{
This function predicts the probability density function (PDF) of a Coxian phase-type model at a given time
for a new observation, based on a trained model object. It uses the current phase and time to adjust the
model's transition rates and calculate the PDF.
}
\examples{
\dontrun{
  # Assuming `model_object` is a result from `trainCoxianModel` and `new_observation` is properly formatted
  pdf_value <- predictCoxianPdf(model_object, new_observation, current_phase = 2, current_time = 10, strata_by = "phase")
  print(pdf_value)
}

}
