% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictProgressiveCoxianModel.R
\name{predictProgressiveCoxianModel}
\alias{predictProgressiveCoxianModel}
\title{Predict Event Times Using Progressive Coxian Model}
\usage{
predictProgressiveCoxianModel(
  model_list,
  new_observations,
  n_phases,
  upper_time = 10000,
  strata_by,
  file_path = NULL
)
}
\arguments{
\item{model_list}{A list of fitted Cox PH model objects, which is the output of `trainProgressiveCoxianModel`.}

\item{new_observations}{A dataframe containing the covariate values for new observations, including the "phase" variable.}

\item{n_phases}{The total number of phases in the Coxian model.}

\item{upper_time}{Optional upper limit for numerical integration to approximate infinity. Defaults to 10000.}

\item{strata_by}{A string specifying the column name in new_observations used for stratification, aligning with the model's structure.}

\item{file_path}{Optional file path to save each row's results to a CSV file. If not provided, results will not be saved to a file.}
}
\value{
The input dataframe with an additional `estimatedEventTime` column containing the expected event time for each observation.
}
\description{
This function predicts the expected event times for new observations using a fitted Progressive Coxian model.
It selects the appropriate model from the list of fitted models based on the "phase" variable in the new observations.
}
\examples{
\dontrun{
  # Assume `model_list` is a list of pre-fitted model objects
  # `new_observations` is a dataframe with multiple rows, each row having covariate values and a "phase" column
  result_df <- predictProgressiveCoxianModel(model_list, new_observations,
                    n_phases = 3, upper_time = 10000, strata_by = "phaseColumn", file_path = "predictions.csv")
  print(result_df)
}

}
