% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressiveCoxianPhModel.R
\name{progressiveCoxianPhModel}
\alias{progressiveCoxianPhModel}
\title{Progressive Coxian Phase Type Model}
\usage{
progressiveCoxianPhModel(
  training_data,
  new_observation,
  n_phases,
  current_phase,
  current_time,
  strata_by
)
}
\arguments{
\item{training_data}{A data frame containing the training data to fit the Cox proportional hazards model.}

\item{new_observation}{A data frame containing the covariate values for the new observation.}

\item{n_phases}{An integer specifying the total number of phases in the Coxian model.}

\item{current_phase}{An integer specifying the current phase of the process (must be between 1 and n_phases).}

\item{current_time}{A numeric value specifying the time at which to calculate the Coxian PDF.}

\item{strata_by}{A string specifying the column name in the data frame to be used for stratification.}
}
\value{
A numeric value representing the calculated PDF of the Coxian phase type model at the specified time.
}
\description{
This function computes the probability density function (PDF) of a Coxian phase type model at a given time,
considering the current phase of the process. It adjusts the number of phases and the data accordingly
based on the current phase before fitting the Cox proportional hazards model and estimating the transition rates.
}
\examples{
\dontrun{
  # Assuming training_data is a data frame with appropriate columns and new_observation is a data frame with the covariate values for the new observation
  coxian_pdf <- progressiveCoxianPhModel(training_data, new_observation, n_phases = 3, current_phase = 1, current_time = 5, strata_by = "phase")
  print(coxian_pdf)
}

}
