% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveModelToFile.R
\name{saveModelToFile}
\alias{saveModelToFile}
\title{Save Model to File}
\usage{
saveModelToFile(model, filepath)
}
\arguments{
\item{model}{The model object to be saved.}

\item{filepath}{A string specifying the path to the file where the model object should be saved. The file
will be created or overwritten using `saveRDS()`.}
}
\value{
NULL. The function is called for its side effects (saving the model to disk). If an error occurs during saving, an error message is printed to the console.
}
\description{
This function saves a model object to a specified file path using `saveRDS()`. This is useful for
scenarios where you need to persist model objects on disk for later use, such as for prediction or further analysis.
}
\examples{
\dontrun{
  # Assuming you have a model object `my_model`
  model_path <- "path/to/your_saved_model.rds"
  saveModelToFile(my_model, model_path)
  # The model is now saved to the specified file path
}

}
